refactor: 3차 피드백 반영
