+++ src/main/java/wooteco/subway/admin/domain/Path.java
@@ -1,17 +1,18 @@
 package wooteco.subway.admin.domain;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class Path {
     private Stations path;
     private int totalWeight;
     private int totalInformation;
+    private PathType pathType;
 
-    public Path(Stations path, int totalWeight, int totalInformation) {
+    public Path(Stations path, int totalWeight, int totalInformation, PathType pathType) {
         this.path = path;
         this.totalWeight = totalWeight;
         this.totalInformation = totalInformation;
+        this.pathType = pathType;
     }
 
     private Path() {
@@ -23,25 +24,33 @@ public class Path {
         SubwayMap subwayMap = SubwayGraph.makeGraph(type, stations, lineStations);
         List<Long> shortestPath = subwayMap.findShortestPath(sourceId, targetId);
 
-        Stations pathStations = shortestPath.stream()
-                .map(stations::findStation)
-                .collect(Collectors.collectingAndThen(Collectors.toList(), Stations::new));
+        Stations pathStations = stations.makePathStations(shortestPath);
 
         int weight = subwayMap.getPathWeight(sourceId, targetId);
         int information = lineStations.getInformation(shortestPath, type);
 
-        return new Path(pathStations, weight, information);
+        return new Path(pathStations, weight, information, type);
     }
 
     public Stations getPath() {
         return path;
     }
 
-    public int getTotalInformation() {
+    public List<Station> getPathStations() {
+        return path.getStations();
+    }
+
+    public int getDistance() {
+        if (this.pathType.equals(PathType.DISTANCE)) {
+            return totalWeight;
+        }
         return totalInformation;
     }
 
-    public int getTotalWeight() {
-        return totalWeight;
+    public int getDuration() {
+        if (this.pathType.equals(PathType.DURATION)) {
+            return totalWeight;
+        }
+        return totalInformation;
     }
 }
