+++ src/main/java/wooteco/subway/admin/domain/Stations.java
@@ -3,6 +3,7 @@ package wooteco.subway.admin.domain;
 import wooteco.subway.admin.domain.exception.NoExistStationException;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class Stations {
     private List<Station> stations;
@@ -11,15 +12,21 @@ public class Stations {
         this.stations = stations;
     }
 
-    public List<Station> getStations() {
-        return stations;
+    public Stations makePathStations(List<Long> shortestPath) {
+        return shortestPath.stream()
+                .map(this::findStation)
+                .collect(Collectors.collectingAndThen(Collectors.toList(), Stations::new));
     }
 
-    public Station findStation(Long id) {
+    private Station findStation(Long id) {
         return stations.stream()
                 .filter(station -> station.getId().equals(id))
                 .findFirst()
                 .orElseThrow(NoExistStationException::new);
+    }
+
+    public List<Station> getStations() {
+        return stations;
     }
 }
