+++ src/test/java/wooteco/subway/admin/acceptance/LineAcceptanceTest.java
@@ -6,7 +6,10 @@ import wooteco.subway.admin.dto.LineDetailResponse;
 import wooteco.subway.admin.dto.LineResponse;
 
 import java.time.LocalTime;
+import java.time.format.DateTimeFormatter;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -46,5 +49,22 @@ public class LineAcceptanceTest extends AcceptanceTest {
         // then
         List<LineResponse> linesAfterDelete = getLines();
         assertThat(linesAfterDelete.size()).isEqualTo(3);
+    }
+
+    List<LineResponse> getLines() {
+        return getList("/lines", LineResponse.class);
+    }
+
+    void updateLine(Long id, LocalTime startTime, LocalTime endTime) {
+        Map<String, String> params = new HashMap<>();
+        params.put("startTime", startTime.format(DateTimeFormatter.ISO_LOCAL_TIME));
+        params.put("endTime", endTime.format(DateTimeFormatter.ISO_LOCAL_TIME));
+        params.put("intervalTime", "10");
+
+        put("/lines/" + id, params);
+    }
+
+    void deleteLine(Long id) {
+        delete("/lines/" + id);
     }
 }
